+++ pages/_campus/phones/_id/edit.vue
@@ -97,6 +97,8 @@
         v-if="id"
         type="submit"
         class="button is-primary"
+        :class="{'is-loading': loading}"
+        :disabled="loading"
       >
         <span class="icon is-small">
           <fa-icon :icon="['fas', 'save']" />
@@ -108,6 +110,8 @@
         v-else
         type="submit"
         class="button is-primary"
+        :class="{'is-loading': loading}"
+        :disabled="loading"
       >
         <span class="icon is-small">
           <fa-icon :icon="['fas', 'plus']" />
@@ -123,6 +127,7 @@ import { mapGetters } from 'vuex';
 import ecField from '~/components/form/field.vue';
 import ecSearchPhoneStates from '~/components/form/search-phone-states.vue';
 import ecSearchPhoneModels from '~/components/form/search-phone-models.vue';
+import toggleLoading from '~/helpers/mixins/toggle-loading';
 
 const EDITABLE_FIELDS = [
   'id',
@@ -140,6 +145,7 @@ export default {
     ecSearchPhoneStates,
     ecSearchPhoneModels,
   },
+  mixins: [toggleLoading],
   props: {
     phone: {
       type: Object,
@@ -163,6 +169,7 @@ export default {
     async edit(phone) {
       let data = {};
       try {
+        this.toggleLoading(true);
         if (this.id) {
           ({ data } = (await this.PhonesAPI.patchPhone(phone.id, phone)));
         } else {
@@ -179,6 +186,7 @@ export default {
           this.$toast.error('Erreur serveur, si le problème persiste, veuillez contacter le service technique');
         }
       }
+      this.toggleLoading(false);
     },
   },
 };